EXPDAT-026 added detail view cancel close test, added test Util class, fixed keyboard visible on open list [MF,WB]
