+++ app/src/main/java/at/sw21_tug/team_25/expirydates/utils/Util.kt
@@ -4,6 +4,8 @@ import android.app.Activity
 import android.content.Context
 import android.content.res.Configuration
 import android.content.res.Resources
+import android.view.View
+import android.view.inputmethod.InputMethodManager
 import android.widget.Toast
 import java.text.SimpleDateFormat
 import java.util.*
@@ -48,5 +50,12 @@ class Util {
             )
             toast.show()
         }
+
+        fun hideKeyboard(activity: Activity, view: View) {
+
+            val imm = activity.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
+            imm.hideSoftInputFromWindow(view.windowToken, 0)
+        }
+
     }
 }
